rename and rewrite `signViaMetaMask` to `signViaEthWallet` using new api
